Move to a metrics class
