+++ app/util/StepFunctions.scala
@@ -5,12 +5,11 @@ import software.amazon.awssdk.services.sfn.model._
 import com.fasterxml.jackson.core.JsonParseException
 import com.gu.media.upload.model._
 import play.api.libs.json.{JsResultException, Json}
-import com.gu.media.telemetry.Telemetry
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.jdk.CollectionConverters._
 
-class StepFunctions(awsConfig: AWSConfig, telemetry: Telemetry) {
+class StepFunctions(awsConfig: AWSConfig) {
   def getById(id: String): Option[Upload] = {
     val arn =
       s"${awsConfig.pipelineArn.replace(":stateMachine:", ":execution:")}:$id"
@@ -62,11 +61,6 @@ class StepFunctions(awsConfig: AWSConfig, telemetry: Telemetry) {
       .stateMachineArn(awsConfig.pipelineArn)
       .input(Json.stringify(Json.toJson(upload)))
       .build()
-//    telemetry.sendTelemetryEvent(
-//      "start",
-//      Map("id" -> upload.id, "parts" -> upload.parts.length.toString)
-//    )
-
     awsConfig.stepFunctionsClient.startExecution(stepFunctionsRequest)
   }
